[server-pro] test: harden reconfiguring (#28193)
When reconfiguring fails, the config snapshot in the frontend is not
updated. When the next test uses the previous config, it will not
attempt to restore that config again, and instead use the other config.
Reset the config before making changes to ensure all tests get their
desired config, even when reconfiguring fails.

GitOrigin-RevId: 0939e709cf93da820a86fef2599940669138d728